total update 22.3.13

+++ Assets/_Scripts/Data/RaceIdleData.cs
@@ -19,7 +19,7 @@ public class RaceIdleData
     public RaceData raceData;
 
 
-    public int Money;
+    public float Money;
 
 
 }
